+++ src/containers/SearchPage/SortBy/SortByPopup.js
@@ -10,19 +10,17 @@ const optionLabel = (options, key) => {
   return option ? option.label : key;
 };
 
-const SortByIcon = () => {
+const SortByIcon = props => {
+  const classes = classNames(css.icon, props.className);
+  // extra small arrow head (down)
   return (
-    <svg className={css.icon} width="10" height="16" xmlns="http://www.w3.org/2000/svg">
-      <g
-        stroke="#4a4a4a"
-        strokeWidth="1.5"
-        fill="none"
+    <svg className={classes} width="8" height="5" xmlns="http://www.w3.org/2000/svg">
+      <path
+        d="M3.764 4.236c.131.13.341.13.472 0l2.666-2.667a.333.333 0 10-.471-.471L4 3.528l-2.43-2.43a.333.333 0 10-.471.471l2.665 2.667z"
+        fill="#4A4A4A"
+        stroke="#4A4A4A"
         fillRule="evenodd"
-        strokeLinecap="round"
-        strokeLinejoin="round"
-      >
-        <path d="M3.25 7.125v7.438M5 12.813l-1.75 1.75-1.75-1.75M6.75 8.875V1.438M5 3.188l1.75-1.75 1.75 1.75" />
-      </g>
+      />
     </svg>
   );
 };
@@ -62,18 +60,20 @@ class SortByPopup extends Component {
 
     const classes = classNames(rootClassName || css.root, className);
     const menuLabelClasses = classNames(menuLabelRootClassName || css.menuLabel);
+    const iconArrowClassName = this.state.isOpen ? css.iconArrowAnimation : null;
 
     return (
       <Menu
         className={classes}
         useArrow={false}
         contentPlacementOffset={contentPlacementOffset}
+        contentPosition="left"
         onToggleActive={this.onToggleActive}
         isOpen={this.state.isOpen}
       >
         <MenuLabel className={menuLabelClasses}>
-          <SortByIcon />
           {menuLabel}
+          <SortByIcon className={iconArrowClassName} />
         </MenuLabel>
         <MenuContent className={css.menuContent}>
           <MenuItem key="sort-by">